Remove tightly coupling of publish task name
- Currently, if you want to make your own publish task,
its name is set to a hardcoded name "publishArtifact".

Typically, this field is set by the gradle environment during
factory construction, and so mutating the state of this creates
inconsistencies where gradle *thinks* a task has one name,
but the task itself does not report to that name.

This conflict prevents the task from being executed.